Create README.md
This GitHub repository hosts the source code for the NODA website, which serves as the final product of the project. The website features a user interface where users can access a database of annotated texts using T-MODA (Technology for Information Transformation based on Object-Document Analysis) and an instructional section on how to utilize T-MODA effectively.